"check_style_score -> 0 of 1 passed. style score: 0/1 "

+++ cspp2-assignments/M4/Assignment-3/Solution.java
@@ -23,7 +23,7 @@ public class Solution
 	public static String binaryToDecimal(String a) {
 		int count = 0;
 		System.out.println(a);
-		long b = Integer.parseInt(a);
+		long b = Long.parseLong(a);
 		long indi = 0;
 		double num = 0;
 		final int ten = 10;
